requirements.txt
phidata
python-dotenv
duckduckgo-search
fastapi
uvicorn
groq
google-generativeai
sqlalchemy
pgvector
psycopg[binary]
pypdf
googlesearch-python 
pycountry
streamlit
bs4
fitz
